refactor: 폴더 이름 변경
- validator 폴더를 utils 폴더로 이름 변경
